requirements.txt
pandas
numpy<2.0
pyyaml
jpype1
scikit-learn
mlflow
pygraphviz
fairlearn
ipykernel
flatten-dict
prdc
pydot
seaborn
umap-learn
git+https://github.com/cmu-phil/py-tetrad
git+https://github.com/py-why/dowhy@main
pymdma
torch>=1.10.0
causalflows>=0.1.0
tqdm>=4.62.3
optuna>=3.0.0